Moving google cloud specific output managers to google_cloud folder
Fixes #22
