+++ transcoder/output/google_cloud/BigQueryOutputManager.py
