+++ transcoder/output/google_cloud/PubSubOutputManager.py
